creates guess input component
